refactor: packages/ui-voip mark app functionality as experimental
