+++ packages/ui-voip/src/providers/MediaCallProvider.tsx
@@ -2,6 +2,9 @@ import type { ReactNode } from 'react';
 
 import MediaCallInstanceProvider from './MediaCallInstanceProvider';
 import MediaCallViewProvider from './MediaCallViewProvider';
+import MediaCallAppActionsContext, {
+	defaultMediaCallAppActionsContextValue,
+} from '../experimental/AppActionButtons/context/MediaCallAppActionsContext';
 
 type MediaCallProviderProps = {
 	children: ReactNode;
@@ -10,8 +13,9 @@ type MediaCallProviderProps = {
 const MediaCallProvider = ({ children }: MediaCallProviderProps) => {
 	return (
 		<MediaCallInstanceProvider>
-			<MediaCallViewProvider />
-			{children}
+			<MediaCallViewProvider>
+				<MediaCallAppActionsContext.Provider value={defaultMediaCallAppActionsContextValue}>{children}</MediaCallAppActionsContext.Provider>
+			</MediaCallViewProvider>
 		</MediaCallInstanceProvider>
 	);
 };
